Add implements

+++ core/src/main/scala/com/github/j5ik2o/dddbase/AggregateMultiWriter.scala
@@ -2,6 +2,6 @@ package com.github.j5ik2o.dddbase
 
 trait AggregateMultiWriter[M[_]] extends AggregateIO[M] {
 
-  def storeMulti(aggregates: Seq[AggregateType]): M[Int]
+  def storeMulti(aggregates: Seq[AggregateType]): M[Long]
 
 }
