+++ core/src/main/scala/com/github/j5ik2o/dddbase/AggregateSingleWriter.scala
@@ -2,6 +2,6 @@ package com.github.j5ik2o.dddbase
 
 trait AggregateSingleWriter[M[_]] extends AggregateIO[M] {
 
-  def store(aggregate: AggregateType): M[Int]
+  def store(aggregate: AggregateType): M[Long]
 
 }
