+++ core/src/main/scala/com/github/j5ik2o/dddbase/AggregateSoftDeletable.scala
@@ -2,6 +2,6 @@ package com.github.j5ik2o.dddbase
 
 trait AggregateSoftDeletable[M[_]] { this: AggregateIO[M] =>
 
-  def softDelete(id: IdType): M[Int]
+  def softDelete(id: IdType): M[Long]
 
 }
